19.9.2025 Version 0.16
Now is possible to create new playlist. However, next step is to load created playlists in "Your Library" screen so that user can open his playlist and most importantly add songs to the playlist. Next objectives:

Added are playlist screen which displays selected playlist and also opens new selectSongs screen which is lets users to add songs to the playlist. Also, it is possible to remove song from the playlist. Next steps are

-> Play playlist
-> Add random play
-> Artist view
-> implement search functionality to selectSongs
-> search screen
-> search functionality on search screen
-> Album/Playlist preview